build: add jenkinsfile and flake file
To be able to deploy site we need a pipeline to build it and publish it
to a branch from which our caddy-git service will pull for changes.
All of this will be done using nix flake to ease out the
build/development process.

Referenced issue: https://github.com/status-im/status-network-docs/issues/1

Signed-off-by: markoburcul <marko@status.im>